refactor!: internalize shared code and prepare for archive
Removes the 'shared' module and duplicates its utility classes into 'maps-rx' and 'places-rx' as internal extensions.

This change eliminates the need for a separate 'shared' Maven artifact, allowing 'maps-rx' and 'places-rx' to be used as standalone libraries without complex transitive dependencies for manual AAR consumers. Code duplication was chosen over a shared artifact to simplify distribution as we prepare to archive these libraries.

Key changes:
- Deleted 'shared' module and updated settings.gradle.
- Copied MainThreadObservable and related utilities to 'com.google.maps.android.rx.maps.internal' and 'com.google.maps.android.rx.places.internal'.
- Updated all imports to use the local internal versions.
- Removed project dependency on ':shared' from build files.
- Marked all duplicated code as 'internal' to prevent binary compatibility issues or duplicate class definitions if both libraries are used together.

This refactoring streamlines the project structure in anticipation of archiving the repository.
