+++ gradle/libs.versions.toml
@@ -3,56 +3,41 @@ compileSdk = "36"
 minSdk = "24"
 targetSdk = "36"
 
+activity = "1.12.2"
+appCompat = "1.7.1"
+dokka-gradle-plugin = "2.1.0"
+gradle = "8.13.2"
+gradleMavenPublishPlugin = "0.35.0"
+jacoco-android = "0.2.1"
 kotlin = "2.3.0"
+lifecycleRuntimeKtx = "2.10.0"
+mapsKtx = "5.2.1"
+material = "1.13.0"
+places = "5.1.1"
+playServicesMaps = "19.2.0"
+rxLifecycle = "4.0.2"
 rxandroid = "3.0.2"
 rxjava = "3.1.12"
-gradle = "8.13.2"
-jacoco-android = "0.2.1"
-dokka-gradle-plugin = "2.1.0"
 secrets-gradle-plugin = "2.0.1"
-playServicesMaps = "19.2.0"
-places = "5.1.1"
-appCompat = "1.7.1"
-activity = "1.12.2"
-lifecycleRuntimeKtx = "2.10.0"
-material = "1.13.0"
-rxLifecycle = "4.0.2"
-mapsKtx = "5.2.1"
 volley = "1.2.1"
-gradleMavenPublishPlugin = "0.35.0"
-junit = "4.13.2"
-mockito = "5.21.0"
-mockitoInline = "5.2.0"
-mockitoKotlin = "6.1.0"
-mockk = "1.14.7"
-robolectric = "4.16"
-androidxTestCore = "1.7.0"
-androidxTestExtJunit = "1.3.0"
 
 [libraries]
-jetbrainsKotlinStdlib = { module = "org.jetbrains.kotlin:kotlin-stdlib", version.ref = "kotlin" }
-kotlinStdlib = { module = "org.jetbrains.kotlin:kotlin-stdlib-jdk8", version.ref = "kotlin" }
-rxJava = { module = "io.reactivex.rxjava3:rxjava", version.ref = "rxjava" }
-rxAndroid = { module = "io.reactivex.rxjava3:rxandroid", version.ref = "rxandroid" }
-playServicesMaps = { module = "com.google.android.gms:play-services-maps", version.ref = "playServicesMaps" }
-places = { module = "com.google.android.libraries.places:places", version.ref = "places" }
-appCompat = { module = "androidx.appcompat:appcompat", version.ref = "appCompat" }
 activityKtx = { module = "androidx.activity:activity-ktx", version.ref = "activity" }
-lifecycleRuntimeKtx = { module = "androidx.lifecycle:lifecycle-runtime-ktx", version.ref = "lifecycleRuntimeKtx" }
-material = { module = "com.google.android.material:material", version.ref = "material" }
-rxLifecycle = { module = "com.trello.rxlifecycle4:rxlifecycle-android-lifecycle-kotlin", version.ref = "rxLifecycle" }
-mapsKtx = { module = "com.google.maps.android:maps-ktx", version.ref = "mapsKtx" }
-volley = { module = "com.android.volley:volley", version.ref = "volley" }
-kotlin-gradle-plugin = { module = "org.jetbrains.kotlin:kotlin-gradle-plugin", version.ref = "kotlin" }
+appCompat = { module = "androidx.appcompat:appcompat", version.ref = "appCompat" }
 dokka-gradle-plugin = { module = "org.jetbrains.dokka:dokka-gradle-plugin", version.ref = "dokka-gradle-plugin" }
-jacoco-android = { module = "com.mxalbert.gradle:jacoco-android", version.ref = "jacoco-android" }
-secrets-gradle-plugin = { module = "com.google.android.libraries.mapsplatform.secrets-gradle-plugin:secrets-gradle-plugin", version.ref = "secrets-gradle-plugin" }
 gradle = { module = "com.android.tools.build:gradle", version.ref = "gradle" }
 gradle-maven-publish-plugin = { module = "com.vanniktech:gradle-maven-publish-plugin", version.ref = "gradleMavenPublishPlugin" }
-junit = { module = "junit:junit", version.ref = "junit" }
-mockitoCore = { module = "org.mockito:mockito-core", version.ref = "mockito" }
-mockitoKotlin = { module = "org.mockito.kotlin:mockito-kotlin", version.ref = "mockitoKotlin" }
-mockk = { module = "io.mockk:mockk", version.ref = "mockk" }
-robolectric = { module = "org.robolectric:robolectric", version.ref = "robolectric" }
-androidxTestCore = { module = "androidx.test:core", version.ref = "androidxTestCore" }
-androidxTestExtJunit = { module = "androidx.test.ext:junit", version.ref = "androidxTestExtJunit" }
+jacoco-android = { module = "com.mxalbert.gradle:jacoco-android", version.ref = "jacoco-android" }
+jetbrainsKotlinStdlib = { module = "org.jetbrains.kotlin:kotlin-stdlib", version.ref = "kotlin" }
+kotlin-gradle-plugin = { module = "org.jetbrains.kotlin:kotlin-gradle-plugin", version.ref = "kotlin" }
+kotlinStdlib = { module = "org.jetbrains.kotlin:kotlin-stdlib-jdk8", version.ref = "kotlin" }
+lifecycleRuntimeKtx = { module = "androidx.lifecycle:lifecycle-runtime-ktx", version.ref = "lifecycleRuntimeKtx" }
+mapsKtx = { module = "com.google.maps.android:maps-ktx", version.ref = "mapsKtx" }
+material = { module = "com.google.android.material:material", version.ref = "material" }
+places = { module = "com.google.android.libraries.places:places", version.ref = "places" }
+playServicesMaps = { module = "com.google.android.gms:play-services-maps", version.ref = "playServicesMaps" }
+rxAndroid = { module = "io.reactivex.rxjava3:rxandroid", version.ref = "rxandroid" }
+rxJava = { module = "io.reactivex.rxjava3:rxjava", version.ref = "rxjava" }
+rxLifecycle = { module = "com.trello.rxlifecycle4:rxlifecycle-android-lifecycle-kotlin", version.ref = "rxLifecycle" }
+secrets-gradle-plugin = { module = "com.google.android.libraries.mapsplatform.secrets-gradle-plugin:secrets-gradle-plugin", version.ref = "secrets-gradle-plugin" }
+volley = { module = "com.android.volley:volley", version.ref = "volley" }
