+++ maps-rx/src/main/java/com/google/maps/android/rx/GoogleMapCameraIdleObservable.kt
@@ -15,7 +15,7 @@
 package com.google.maps.android.rx
 
 import com.google.android.gms.maps.GoogleMap
-import com.google.maps.android.rx.shared.MainThreadObservable
+import com.google.maps.android.rx.maps.internal.MainThreadObservable
 import io.reactivex.rxjava3.android.MainThreadDisposable
 import io.reactivex.rxjava3.core.Observable
 import io.reactivex.rxjava3.core.Observer
