+++ places-rx/src/main/java/com/google/maps/android/rx/places/internal/NotOnMainThreadException.kt
@@ -1,0 +1,24 @@
+// Copyright 2021 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package com.google.maps.android.rx.places.internal
+
+import kotlin.IllegalStateException
+
+/**
+ * Exception for when an action does not occur on the Android main thread when it should.
+ */
+internal class NotOnMainThreadException : IllegalStateException(
+    "Subscription is on ${Thread.currentThread().name} but must be on main thread."
+)
