+++ settings.gradle
@@ -1,6 +1,6 @@
 include ':app'
 include ':maps-rx'
-include ':shared'
+
 rootProject.name = "android-maps-rx"
 include ':places-rx'
 include ':docs'
